+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Version 3.0.11
-- Drag Ruler units now match the grid units of the scene. [#225](https://github.com/dmdorman/hero6e-foundryvtt/issues/225)
+- Drag Ruler units now match grid units of the scene. [#225](https://github.com/dmdorman/hero6e-foundryvtt/issues/225)
+- POWER DEFENSE works vs DRAIN.
+- DELAYED RETURN RATE works vs AID/DRAIN.
 
 # Version 3.0.10
 - Temporary changes to CHARACTERISTIC MAX have red/green backgrounds on character sheet, similar to how VALUE background turns red/green.